airu_flask/requirements.txt
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
blinker==1.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
firebase-admin==3.0.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
gcloud==0.18.3
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.17.0
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jws==0.1.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
mistune==0.8.4
msgpack==0.6.1
Naked==0.1.31
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
oauth2client==4.1.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodome==3.9.0
Pygments==2.4.2
pyparsing==2.4.2
Pyrebase4==4.3.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-jwt==2.0.1
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
rebase==1.3.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
Send2Trash==1.5.0
shellescape==3.4.1
simplejson==3.16.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
WTForms==2.2.1